Remove javadoc linting
JAXB generates Javadocs with invalid HTML and the Java 8 linter fails
Javadoc generation because of this. The Java 7 Javadoc linter wasn't as
strict, so this wasn't a problem before.

We'll be removing JAXB when we migrate from Java 8, so removing the
linting as a temporary fix doesn't feel too bad.